refactor(src) for submit specification
