+++ .vscode/c_cpp_properties.json
@@ -3,7 +3,7 @@
         {
             "name": "Linux",
             "includePath": [
-                "${workspaceFolder}/**"
+                "${workspaceFolder}/include"
             ],
             "defines": [],
             "intelliSenseMode": "linux-gcc-x64",
