+++ basic/rbtree/Makefile
@@ -1,0 +1,8 @@
+.PHONY: clean
+
+CFLAGS=-Wall -g
+
+driver: driver.o rbtree.o
+
+clean:
+	rm -f driver *.o
