+++ basic/rbtree/stack.c
