+++ src/driver.c
@@ -1,0 +1,8 @@
+#include <assert.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+
+#include "rbtree.h"
+
+int main(int argc, char *argv[]) { return 0; }
